tests/test-requirements.txt
testtools
python-heatclient
python-muranoclient
python-keystoneclient
requests
